refactor(inference): update docs, keep hooks under the same api

+++ autocrit/inference/inference_hook.py
@@ -1,51 +1,64 @@
-import signal
-import aiohttp
+import argparse
 import asyncio
 import json
-import subprocess
-import torch
-import time
-from tqdm.asyncio import tqdm_asyncio
-
-from abc import ABC, abstractmethod
-import argparse
+import logging
 import os
+import signal
+import subprocess
+import sys
+import time
+from abc import ABC, abstractmethod
+from typing import Any, Dict, List, Optional, Tuple
+
+import aiohttp
 import torch
 import transformers
-from typing import Tuple, Any, Optional, List, Dict
 import tritonclient.grpc.aio as grpcclient
-from autocrit.inference.utils import triton_call, best_of_n
+from autocrit.inference.utils import best_of_n, triton_call
 from text_generation import Client
-import logging
-
-import signal
-import sys
+from tqdm.asyncio import tqdm_asyncio
 
 
 class InferenceHook(ABC):
     def __init__(self, **kwargs):
         """
-        kwargs: a dictionary of parameters to initilize the model
+        Args:
+            kwargs (`Dict[str, Any]`): a dictionary of parameters to initilize the model with
         """
         pass
 
     @abstractmethod
-    def generate(self, prompts: List[str], **kwargs: Dict[str, Any]):
+    def generate(self, prompts: List[str], **kwargs: Dict[str, Any]) -> List[Dict[str, Any]]:
         """
-        prompts: inputs for generations
-        kwargs: parameters to control generation
+        Args:
+            prompts (`List[str]`): inputs for generations
+            kwargs (`Dict[str, Any]`): parameters to control generation
+
+        Returns:
+            outputs (`List[Dict[str, Any]]`): a list of dictionaries, each dictionary contains the following keys:
+                id (`int`): the id of the prompt
+                prompt (`str`): the prompt
+                outputs (`List[str]`): a list of outputs per prompt
         """
         pass
 
     @abstractmethod
     def free(self):
         """
-        Clean up resources after inference
+        Clean up resources after the inference
         """
         pass
 
 class vLLMHook(InferenceHook):
     def __init__(self, model_path, tensor_parallel_size=1, external=False, num_nodes=1):
+        """
+
+        Args:
+            model_path (`str`): the path to the model
+            tensor_parallel_size (`int`): the number of GPUs to use per one server
+            external (`bool`): whether to spawn a new slurm jobs and generate on external nodes or use the current node
+            num_nodes (`int`): the number of nodes to use if external is True
+        """
         self.init_time = time.time()
         self.model_path = model_path
         self.tensor_parallel_size = tensor_parallel_size
@@ -63,8 +76,7 @@ class vLLMHook(InferenceHook):
             sbatch_script_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "vllm.sbatch")
             for _ in range(num_nodes):
                 cmd = f"sbatch {sbatch_script_path} NUM_TP={tensor_parallel_size} MODEL_PATH={model_path} DEVICES={'|'.join(devices)}"
-                print(f'{cmd=}')
-                process = subprocess.Popen(cmd.split(), stdout=subprocess.PIPE, stderr=subprocess.PIPE, env={**os.environ, "TORCHELASTIC_USE_AGENT_STORE": ""})
+                process = subprocess.Popen(cmd.split(), env={**os.environ, "TORCHELASTIC_USE_AGENT_STORE": ""})
 
                 while True:
                     output = process.stdout.readline().decode("utf-8").strip()
@@ -94,7 +106,12 @@ class vLLMHook(InferenceHook):
             self.processes = []
             for i in range(self.data_parallel_size):
                 cmd = f"python -m vllm.entrypoints.api_server -tp={tensor_parallel_size} --model={model_path} --port {8000+i}"
-                process = subprocess.Popen(cmd.split(), stdout=subprocess.PIPE, stderr=subprocess.PIPE, env={**os.environ, "CUDA_VISIBLE_DEVICES": devices[i], "TORCHELASTIC_USE_AGENT_STORE": ""})
+                kwargs = {"env": {**os.environ, "CUDA_VISIBLE_DEVICES": devices[i], "TORCHELASTIC_USE_AGENT_STORE": ""}}
+                if not os.environ.get("DEBUG", False):
+                    kwargs["stdout"] = subprocess.DEVNULL
+                    kwargs["stderr"] = subprocess.DEVNULL
+
+                process = subprocess.Popen(cmd.split(), **kwargs)
                 self.processes.append(process)
 
             print(f"Loading {self.data_parallel_size} processes for {model_path}...")
@@ -138,7 +155,7 @@ class vLLMHook(InferenceHook):
                     return {"id": i, "prompt": prompt, "outputs": [None]}
 
 
-    def generate(self, prompts, **kwargs):
+    def generate(self, prompts: List[str], **kwargs) -> List[Dict[str, Any]]:
         async def generate_vllm_api(prompts, **kwargs):
             outputs = [self.request_vllm_api(prompt=prompt, i=i, **kwargs) for i, prompt in enumerate(prompts)]
             return await tqdm_asyncio.gather(*outputs, desc=f"Inferencing {self.model_path}")
@@ -152,19 +169,30 @@ class vLLMHook(InferenceHook):
 
     def free(self):
         if self.external:
-            subprocess.run(f"scancel {' '.join(self.job_ids)}".split())
+            if self.job_ids:
+                subprocess.run(f"scancel {' '.join(self.job_ids)}".split())
+                self.job_ids = []
         else:
             for p in self.processes:
                 os.kill(p.pid, signal.SIGTERM)
                 p.communicate()
             print(f"Unloaded all {self.model_path} processes")
+            self.processes = []
 
     def __del__(self):
         self.free()
 
-# Inference hook that uses the HuggingFace API to call a model
 class HuggingFaceHook(InferenceHook):
+    """
+    Inference hook that uses plain HuggingFace transformers API
+    """
+
     def __init__(self, model_path: str, tokenizer_path : Optional[str] = None):
+        """
+        Args:
+            model_path (`str`): the directory of the model
+            tokenizer_path (`str`): the directory of the tokenizer, if None, the `model_path` is implied
+        """
         self.model = transformers.AutoModelForCausalLM.from_pretrained(model_path, device_map="auto")
         self.tokenizer = transformers.AutoTokenizer.from_pretrained(tokenizer_path or model_path)
         if self.tokenizer.pad_token is None:
@@ -199,101 +227,89 @@ class HuggingFaceHook(InferenceHook):
         del self.model
         del self.tokenizer
 
-# Inference hook that uses the HuggingFace API to call a model. Uses the best of N sampling method
 class HuggingFaceHookBestOfN(HuggingFaceHook):
-    def __init__(self, dir : str, tokenizer_name : Optional[str] = None):
-        """
-        dir: the directory of the model
-        tokenizer_name: the name of the tokenizer to use, if None, use the model name
-        """
-        super().__init__(dir, tokenizer_name)
+    """
+    Inference hook that uses the HuggingFace API to call a model. Uses the best of N sampling method
+    """
 
-    def infer(self, input_texts : List[str],
-              generate_params : Dict[str, Any],
-              **kwargs: Any) -> Any:
+    def __init__(self, model_path: str, tokenizer_path : Optional[str] = None):
         """
-        input_texts: a list of strings, each string is a prompt
-        generate_params: a dictionary of parameters to pass to the generate function
-        kwargs: any additional arguments to pass to the generate function
-        returns: a list of strings, each string is a generated output
+        Args:
+            model_path (`str`): the directory of the model
+            tokenizer_path (`str`): the directory of the tokenizer, if None, the `model_path` is implied
+        """
+        super().__init__(model_path, tokenizer_path)
+
+    def generate(self, prompts: List[str], **kwargs: Any) -> List[str]:
+        """
+        Args:
+            prompts: a list of prompts to generate
+            kwargs: a dictionary of parameters to pass to the generate function
         """
 
-        output_txt = best_of_n(self.model, self.tokenizer, input_texts, gen_kwargs=generate_params, **kwargs)
+        output_txt = best_of_n(self.model, self.tokenizer, prompts, gen_kwargs=kwargs)
         return output_txt
 
 
 class TritonHook(InferenceHook):
-    def __init__(self, dir : str, model_name : str, tokenizer_name : Optional[str] = None):
+    def __init__(self, url: str, model_path: str, tokenizer_path : Optional[str] = None):
         """
-        dir: location of the triton server
-        model_name: the name of the model to use
-        tokenizer_name: the name of the tokenizer to use, if None, use the model name
+        Args:
+            url (`str`): location of the triton server
+            model_path (`str`): the name of the model to use
+            tokenizer_path (`str`): the name of the tokenizer to use, if None, the `model_path` is implied
         """
-        super().__init__(dir)
-        self.url = dir # url contains host:port
+        self.url = url # url contains host:port
         # TODO: if URL is a path to a triton model, we shold load the model and launch the server
-        self.model_name = model_name
-        if tokenizer_name is None:
-            self.tokenizer_name = model_name
+        self.model_path = model_path
+        if tokenizer_path is None:
+            self.tokenizer_path = model_path
         else:
-            self.tokenizer_name = tokenizer_name
+            self.tokenizer_path = tokenizer_path
 
-        self.client = None
-        self.tokenizer = None
-
-    def load(self, **kwargs):
         # create a client using url
         self.client = grpcclient.InferenceServerClient(url=self.url)
-        self.tokenizer = transformers.AutoTokenizer.from_pretrained(self.tokenizer_name)
+        self.tokenizer = transformers.AutoTokenizer.from_pretrained(self.tokenizer_path)
         # check if there is a padding token, if not add one
         if self.tokenizer.pad_token is None:
             self.tokenizer.add_special_tokens({"pad_token": "[PAD]"})
 
 
-    def infer(self, input_texts : List[str],
-              generate_params : Dict[str, Any],
-              **kwargs: Any) -> Any:
+    def generate(self, prompts: List[str], **kwargs: Any) -> List[str]:
         """
-        input_texts: a list of strings, each string is a prompt
-        generate_params: a dictionary of parameters to pass to the generate function
-        kwargs: any additional arguments to pass to the generate function
-        returns: a list of strings, each string is a generated output
+        Args:
+            prompts: a list of prompts to generate
+            kwargs: a dictionary of parameters to pass to the generate function
         """
 
         # use self.client to call the model
         # assume input and output are batched
-        inp_text = input_texts
-        inps = self.tokenizer(inp_text, return_tensors="pt", padding=True)
+        inps = self.tokenizer(prompts, return_tensors="pt", padding=True)
         # call infer
-        logits, output_txt = triton_call(self.client, self.model_name, inps.input_ids, **generate_params)
+        logits, output_txt = triton_call(self.client, self.model_path, inps.input_ids, **kwargs)
 
         if not "no_decode" in kwargs:
             output_txt = self.tokenizer.batch_decode(output_txt, skip_special_tokens=True)
 
         # check if logits are needed
-        if generate_params["return_logits"]:
+        if kwargs["return_logits"]:
             return output_txt, logits
         else:
             return output_txt
 
 
-# Inference hook that uses the HuggingFace API to call a model
 class TextGenerationHook(InferenceHook):
-    def __init__(self, dir : str):
-        super().__init__(dir)
-        self.model_name = dir
-        self.client = None
-
-    def load(self, **kwargs):
+    def __init__(self, model_path: str):
+        self.model_path = model_path
         # get num shards and port, used for the launcher script
         num_shards = kwargs.get("num_shards", 1)
         port = kwargs.get("port", 8080)
 
         # check if model name is a URL
-        if not self.model_name.startswith("http"):
+        if not self.model_path.startswith("http"):
             # launch the model using the model name and text_generation_launcher.sh
             # The following line runs the launcher script
-            output = subprocess.run(["sbatch./launch.sbatch MODEL_NAME="+str(self.model_name) + " NUM_SHARD="+str(num_shards) + " PORT="+str(port)], capture_output=True)
+            output = subprocess.run(["sbatch./launch.sbatch MODEL_NAME="+str(self.model_path) + " NUM_SHARD="+str(num_shards) + " PORT="+str(port)], capture_output=True)
             logging.info(output.stdout.decode("utf-8"))
             # check return code
             if output.returncode != 0:
@@ -317,22 +333,18 @@ class TextGenerationHook(InferenceHook):
             # Create the client
             self.client = Client(f"http://{ip}:{port}")
         else:
-            self.client = Client(self.model_name)
+            self.client = Client(self.model_path)
 
-    def infer(self, input_texts : List[str],
-              generate_params : Dict[str, Any],
-              **kwargs: Any) -> Any:
+    def generate(self, prompts: List[str], **kwargs: Dict[str, Any]) -> List[str]:
         """
-        input_texts: a list of strings, each string is a prompt
-        generate_params: a dictionary of parameters to pass to the generate function
-        kwargs: any additional arguments to pass to the generate function
-        returns: a list of strings, each string is a generated output
+        Args:
+            prompts (`List[str]`): a list of strings, each string is a prompt
+            kwargs (`Dict[str, Any]`): a dictionary of parameters to pass to the generate function
         """
         # if input_texts is a list, convert it to a string
         if isinstance(input_texts, list):
             input_texts = input_texts[0]
 
         # use self.client to call the model
-        output_txt = self.client.generate(input_texts, **generate_params).generated_text
-        #print(output_txt)
+        output_txt = self.client.generate(input_texts, **kwargs).generated_text
         return output_txt
